Command pattern applied to the modes of application of Param_Est
Currently, the Param_Est object could be used in four possible ways:
1) parity plot mode
2) excel postprocessing mode
3) parameter optimization mode
4) statistics mode

Each of the modes called different functions of the Receiver class,
ie Param_Est.

However, in the future it is possible that new modes will be created,
and that new features will be added.

Therefore, the command pattern is applied to the invocation of the
methods of Param_Est that perform the actual feature.

An interface Command requires a specific feature (eg "Optimize
Parameters") to implement the execute() method

All (currently 4) implementations of the Command are added, including
a 5th Null object command, called NoCommand

The invoker class, ParamEstimationInvoker contains an array of
Commands that can be set by specifying an integer slot number, that
corresponds to the desired mode. By default, the constructor adds the
Null object to each of the Command array slots.

ParameterEstimationDriver serves as the Loader class that creates
the invoker object, and the specific implementations of the modes.