Encapsulation of reactor input parser
the parser that reads in an excel file with structure info regarding
PFR reactor input is encapsulated in a type called PFRReactorInputParser
interfaced to ReactorInputParsable that requires three methods "parse",
"write" and "read".

Every reactor input, be it for a PFR, CSTR, Flame speed simulation, etc
is now represented by the supertype ReactorInput, which is subtyped for
PFR's by the type called PFRReactorInput.

the reading and writing in the parser is completely separated now.

The second parser is removed.