Use of Decorator pattern for chemkin executables
An superclass for all chemkin routines "AbstractChemkinRoutine"
contains attributes common for all routines such as reaction directory,etc

An superclass for decorators "ChemkinRoutineDecorator" has the
method "getKeyword" in which the keywords are called, specific to
a routine.

All of the chemkin routines are then put into classes such as PFRDecorator
etc.

The type ChemkinRoutines is removed.