First Simulation refactored using Decorator pattern
the first simulation needs to be finished before the other simulations
can be parallelized. This is because CKSolnList needs to be created
and adapted.

therefore, a decorator pattern for CKEmulation was created. A superclass
"AbstractCKEmulation" now extends "Thread" and all attributes
are stored in that type.

An abstract CKemulation decorator type "CKemulationDecorator" contains
the run method of "Thread".

FirstSimulationDecorator subclasses this type.