ModelValues refactored using factory pattern
ModelValues type is removed because it was not maintainable.
Instead, a supertype "ModelValue" is created that has an enum with
three possibilities. The subclasses of ModelValue specify which type
of reactor model value is aimed at. More so, for each ModelValue type
a common but specific method "setValue" to retrieve this info from
the .ckcsv file is defined.

A factory type "ModelValueFactory" is created in which the flag from
the ReactorSetupInput is read from the input.xml file. An implementation
of the ModelValue supertype is then created.

Signed-off-by: Nick Vandewiele <nickvandewiele@gmail.com>