Residual encapsulated
the abstract type "Residual" is supertype for the three
types of simulations currently supported.

because EffluentResidual consists of a multiple value, an iterator was
created that iterates over all the values of the speciesFractions
attribute

a method getSSQValue returns the squared residual or the sum of squared
residuals.

a compute() method is called in all of the constructors and populates
the data attribute.