ReactorType removed
The reactor type that should be called from the binary directory
will not be read from the reactor input file anymore.

Instead, this will be specified in the input.xml file and we trust
the user to create the correct reactor input file according to the reactor
type one specified in the xml input.