create a separate CKSolnList for each simulation
Earlier, CKSolnList was only created in the first simulation, copied
to the working dir, and then copied back to each of the other temp dirs.

Now, since each simulation can have a different CKSolnList because of
sensitivity flags for one particular species that was carried out in only
one simulation, this recipe was ready for disaster.

We now create a CKSolnList in each simulation by decorating the
chemkinRoutine in the RegularSimulationDecorator with a
CreateSolnListDecorator.

The method that modifies the CKSolnList is also moved that class.
The FirstSimulationDecorator is removed since redundant.

CKPackager makes no exception for the first simulation any more and
should therefore be faster.